types: rename types.d.ts to types.ts
